requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Motor Claim Estimator — Python Dependencies
# ─────────────────────────────────────────────────────────────────────────────
#
# Install with:
#   pip install -r requirements.txt
#
# ─────────────────────────────────────────────────────────────────────────────

# Web UI
streamlit>=1.32.0

# Computer Vision & Image Processing
opencv-python>=4.9.0
Pillow>=10.2.0

# YOLOv8 (includes PyTorch dependency resolution)
ultralytics>=8.1.0

# PDF Report Generation
reportlab>=4.1.0

# Data & Numerics
numpy>=1.26.0

# (PyTorch is installed automatically by ultralytics;
#  for CPU-only machines no separate torch install is needed)
